\name{crwbmetareg-package}
\alias{crwbmetareg-package}
\docType{package}
\title{
Cluster Robust Wild Bootstrap Meta Regression.
}
\description{
In meta regression sometimes the studies have multiple effects that are correlated.
For this reason cluster robust standard errors must be computed. However,
since the clusters are unbalanced the wild bootstrap is suggested.
}
\details{
\tabular{lll}{
Package: \tab crwbmetareg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-10-18\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Chatzimichael K., Daskalaki C., Emvalomatis G., Tsagris M. and Vangelis Tzouvelekas V. (2023).
Factors Shaping Innovative Behavior: A Meta-Analysis of Technology Adoption Studies in Agriculture.
https://economics.soc.uoc.gr/el/market/998/factors-shaping-farmers-innovative-behavior-a-meta-analysis-of-technology-adoption-studies-in-agriculture

Oczkowski, E. and Doucouliagos, H. (2015).
Wine prices and quality ratings: a meta-regression analysis.
American Journal of Agricultural Economics, 97(1): 103-121.

Cameron, A. C., Gelbach, J. B. and Miller, D. L. (2008).
Bootstrap-based improvements for inference with clustered errors.
The Review of Economics and Statistics, 90(3): 414-427.
}



